ENTER_KEY maybe submits form, maybe doesn't
either way, try several different ways just to ensure it works.
My belief is it doesn't work, but I've had it error out on trying
to find the element to input RETURN on the next line, so maybe it
did work?
Finally, .submit() of course is the best way, so long as I have the
ID of the actual form