add browser class that waits for element to show
the goals of this browser is to be used in place of the standard
Firefox() browser. It'll wait up to a second for the element to
appear before returning it, which helped a lot in solving the tests'
flakiness. (often it was failing after submitting the folder form
because it was trying to click the folder element before the backend
had responded!